backend/requirements.txt
Django>=5.2,<5.3
coverage
pyaml
django-recurrence>=1.14
python-decouple>=3.8
pandas>=2.2
djangorestframework>=3.15
django-cors-headers>=4.4
